requirements.txt
langgraph>=0.2.0
openai>=1.40.0
pydantic>=2.7.0
playwright>=1.46.0
pillow>=10.4.0
python-dotenv>=1.0.1
